# server/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Antimony API",
        "contact": {
            "name": "Institute for Networking at OST",
            "url": "https://www.ost.ch/en/research-and-consulting-services/computer-science/ins-institute-for-network-and-security",
            "email": "antimony@network.garden"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/users/login/native": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Authenticate via native login",
                "parameters": [
                    {
                        "description": "The native credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CredentialsIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "The provided credentials were invalid"
                    },
                    "401": {
                        "description": "Authentication via native login is disabled"
                    }
                }
            }
        },
        "/users/login/openid": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Authenticate via OpenID provider. Redirects the client to the OpenID provider page.",
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "401": {
                        "description": "Authentication via OpenID is disabled"
                    }
                }
            }
        },
        "/users/login/refresh": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Refresh the access token",
                "responses": {
                    "200": {
                        "description": "The authentication config of the server",
                        "schema": {
                            "$ref": "#/definitions/utils.OkResponse-auth_AuthConfig"
                        }
                    },
                    "401": {
                        "description": "The auth token cookie is not set"
                    },
                    "403": {
                        "description": "The provided auth token was invalid"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.AuthConfig": {
            "type": "object",
            "properties": {
                "native": {
                    "$ref": "#/definitions/auth.NativeAuthConfig"
                },
                "openId": {
                    "$ref": "#/definitions/auth.OpenIdAuthConfig"
                }
            }
        },
        "auth.NativeAuthConfig": {
            "type": "object",
            "properties": {
                "allowEmpty": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                }
            }
        },
        "auth.OpenIdAuthConfig": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                }
            }
        },
...
        "user.CredentialsIn": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "utils.OkResponse-auth_AuthConfig": {
            "type": "object",
            "properties": {
                "payload": {
                    "$ref": "#/definitions/auth.AuthConfig"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}